+++ src/components/academy/index.tsx
@@ -1,23 +1,22 @@
 /* tslint:disable: jsx-no-lambda */
-
 import * as qs from 'query-string'
 import * as React from 'react'
 import { Redirect, Route, RouteComponentProps, Switch } from 'react-router'
 
+import Grading from '../../containers/academy/grading'
 import AssessmentListingContainer from '../../containers/assessment/AssessmentListingContainer'
 import Game from '../../containers/GameContainer'
 import { isAcademyRe } from '../../reducers/session'
 import { HistoryHelper } from '../../utils/history'
 import { assessmentCategoryLink } from '../../utils/paramParseHelpers'
 import { AssessmentCategories, AssessmentCategory } from '../assessment/assessmentShape'
-import Grading from './grading'
 import AcademyNavigationBar from './NavigationBar'
 
 interface IAcademyProps extends IDispatchProps, IOwnProps, IStateProps, RouteComponentProps<{}> {}
 
 export interface IDispatchProps {
-  fetchTokens: (ivleToken: string) => void
-  fetchUsername: () => void
+  handleFetchTokens: (ivleToken: string) => void
+  handleFetchUsername: () => void
 }
 
 export interface IOwnProps {
@@ -72,8 +71,8 @@ export const Academy: React.SFC<IAcademyProps> = props => (
 const checkLoggedIn = (props: IAcademyProps) => {
   const ivleToken = qs.parse(props.location.search).token
   if (ivleToken !== undefined) {
-    props.fetchTokens(ivleToken) // just received a callback from IVLE
-    props.fetchUsername()
+    props.handleFetchTokens(ivleToken) // just received a callback from IVLE
+    props.handleFetchUsername()
     return
   } else if (props.accessToken === undefined) {
     return <Route component={redirectToLogin} />
